requirements.txt
channels==3.0.4
channels_redis==3.3.1
daphne==3.0.2
decorator==5.1.1
Django==4.2.17
django-cors-headers==4.3.0
django-heroku==0.3.1
django-redis==5.3.0
django-webpack-loader==2.0.1
djangorestframework==3.15.2
Jinja2==3.1.4
matplotlib==3.8.0
numpy==1.26.1
pandas==2.1.1
py==1.11.0
requests==2.32.2
retry==0.9.2
scikit-learn==1.5.0
six==1.16.0

--find-links https://download.pytorch.org/whl/torch_stable.html

torch==2.2.0+cpu
whitenoise==6.6.0
